fix: join threads to trigger view refresh.
In theory, this will cause the GUI to be blocked waiting for the tartex
thread. In practice, this is typically less than a fraction of a second
and does not affect usability. We use this as a last resort when all
other thread non-blocking attempts to refresh the view — and have
notifications and error dialogs show up immediately — have failed.