fix: try refreshing dir list from tartex thread.
Drop blocking `thread.join()`, since this causes visible UI freeze when
compiling long-ish LaTeX projects.

Instead, try refreshing the dir contents from the tartex thread
directly. This does not yet quite work, esp. in the case where archiving
fails and there is no new content added to the dir. In this case,
refreshing the dir manually (F5) is needed to trigger the displaying of
the error dialog. This could be — indeed, should be — handled better.

In the successful case, at least a new tar file is added to the dir and
nautilus automatically triggers a refresh, perhaps delayed by a second or so.