feat: persistent notification when compiling.
Make the notification shown during compiling, which can take awhile for
more complex LaTeX projects, persistent. When compilation is done,
replace this persistent notification with non-persistent status
notification. Notify urgency "CRITICAL" is used to make the notification
persistent; there does not seem to be a better way of doing this with
libnotify itself.