fix: set env var COLUMNS for consistent line-wraps
When unset, as is the case when nautilus is launched from the desktop
menus, it is otherwise assumed to be 80, leading to a lot of
not-too-long warning lines getting wrapped by tartex's Rich log handler.
On the other hand, the wrapping becomes dependent on the console width
if nautilus is launched from a terminal emulator.

Setting the 'COLUMNS' env variable to 132 fixes this inconsistency and
avoids wrapping on most lines.