fix: key-press anywhere in dialog to start search
Previously the user had to click on the text-view widget to make it
active, before being able to type to search. Use the entire dialog
window instead to drop the requirement for this additional clicking,
allowing the user to start searching immediately when the dialog
launches by entering key presses directly.