feat: show toast if logfile cannot be found.
In case of latexmk errors, if the log file 'tartex_compile_error.log' is
deleted after the dialog is launched, the "Open Log" button will no
longer work and the "clicked" method will raise an exception. Catch the
exception, use its message to show a toast overlay and move on. This is
a very specific edge case, and it is likely no user will ever see this
toast anyway.