fixup! fix: require nautilus version 4.0 instead of 4.1
We need to explicitly support importing both 4.0 and 4.1 as gi does not
allow specifying minimum required versions. Importing will fail if a
nautilus version "4.1" has been imported prior to nautilus-tartex
being loaded and will raise a `ValueError`. In this case, catch the
exception and try loading version "4.1" for our application as well.