fix: use nautilus window active at menu trigger
The active nautilus window may have changed between clicking to activate the
extension and when compilation completes with error or success.

In the error case, the error dialog box was being tethered to the window
determined to be active "at the present time", which may not be the one
which the user initially clicked on. Fixed by this commit, by ensuring
a reference to the active window is obtained when the extension is
triggered from the context menu and is then propagated and used
throughout the code.