feat: launch app with tar selected as notif action
Refactor to allow updating notification action with different targets.
Set initial target for the action as the working project dir, but update
it to the generated tarball if archiving is successful. In the action
handler, check if the target is a dir and launch nautilus at the dir if
so, but launch nautilus with target selected if it is not (e.g. final
output tarball).

When tartex archiving is successful, clicking on notification takes the
user to the project dir with the output tarball selected, while if
archiving is running or has failed simply opens the project dir.

Replace `app.open` with `Gio.Subprocess` to allow passing cmdline
options to nautilus.