fix: drop custom default-action on notification
The default action uses a SimpleAction registered with the app to do its
thing. However, this only works when the app (nautilus) is still
running.  When the app disappears (last window closed), the registered
action disappears with it and clicking on the notification thereafter
tries to trigger the no-longer existing "open-target" action with an
obvious error.  Ultimately, this means no app window is launched, which
is worse than the default no-action event of launching nautilus at HOME
dir.